refactor(schematics): split monolithic test file into focused per-migration test files
Replace migration.vitest.ts and regex-patterns.vitest.ts with three focused
test files under ng-update/tests/:

  import-migration.vitest.ts  — module path replacements, symbol renames,
                                 removed symbols, sibling preservation, and
                                 regex word-boundary pattern tests.
  template-migration.vitest.ts — output/input bindings, cds-icon scoping,
                                  global attributes, header classes, inline
                                  templates, multi-line scenarios, and
                                  regex boundary tests.
  css-migration.vitest.ts      — property renames/removals, selector renames,
                                  attribute value renames, SCSS mixin removal,
                                  inline styles, false-positive guards, and
                                  regex boundary tests.
  test-helpers.ts              — shared runMigrations() / applyTsTransforms()
                                  used by all three files.

Each file combines schematic-level integration tests (HostTree) with
pattern-level boundary tests (pure transform functions) for the same
migration domain, eliminating the artificial split between the two old files.

Made-with: Cursor
